requirements.txt
boto3==1.24.61
botocore==1.27.61
colorthief==0.2.1
jmespath==1.0.1
numpy==1.23.2
opencv-contrib-python==4.6.0.66
Pillow==9.2.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
